undo x264 changes before reapply
