+++ robot/web/web_notc.go
@@ -4,15 +4,13 @@ package web
 
 import (
 	"context"
-	"math"
 	"net/http"
-	"runtime"
 	"sync"
 
-	"go.viam.com/rdk/gostream"
 	"go.viam.com/rdk/logging"
 	"go.viam.com/rdk/resource"
 	"go.viam.com/rdk/robot"
+	weboptions "go.viam.com/rdk/robot/web/options"
 	"go.viam.com/utils/rpc"
 )
 
@@ -23,13 +21,12 @@ func New(r robot.Robot, logger logging.Logger, opts ...Option) Service {
 		opt.apply(&wOpts)
 	}
 	webSvc := &webService{
-		Named:        InternalServiceName.AsNamed(),
-		r:            r,
-		logger:       logger,
-		rpcServer:    nil,
-		services:     map[resource.API]resource.APIResourceCollection[resource.Resource]{},
-		opts:         wOpts,
-		videoSources: map[string]gostream.HotSwappableVideoSource{},
+		Named:     InternalServiceName.AsNamed(),
+		r:         r,
+		logger:    logger,
+		rpcServer: nil,
+		services:  map[resource.API]resource.APIResourceCollection[resource.Resource]{},
+		opts:      wOpts,
 	}
 	return webSvc
 }
@@ -37,23 +34,20 @@ func New(r robot.Robot, logger logging.Logger, opts ...Option) Service {
 type webService struct {
 	resource.Named
 
-	mu           sync.Mutex
-	r            robot.Robot
-	rpcServer    rpc.Server
-	modServer    rpc.Server
-	streamServer *StreamServer
-	services     map[resource.API]resource.APIResourceCollection[resource.Resource]
-	opts         options
-	addr         string
-	modAddr      string
-	logger       logging.Logger
-	cancelCtx    context.Context
-	cancelFunc   func()
-	isRunning    bool
-	webWorkers   sync.WaitGroup
-	modWorkers   sync.WaitGroup
-
-	videoSources map[string]gostream.HotSwappableVideoSource
+	mu         sync.Mutex
+	r          robot.Robot
+	rpcServer  rpc.Server
+	modServer  rpc.Server
+	services   map[resource.API]resource.APIResourceCollection[resource.Resource]
+	opts       options
+	addr       string
+	modAddr    string
+	logger     logging.Logger
+	cancelCtx  context.Context
+	cancelFunc func()
+	isRunning  bool
+	webWorkers sync.WaitGroup
+	modWorkers sync.WaitGroup
 }
 
 // Update updates the web service when the robot has changed.
@@ -66,74 +60,16 @@ func (svc *webService) Reconfigure(ctx context.Context, deps resource.Dependenci
 	return nil
 }
 
+// stub implementation when gostream not available
+func (svc *webService) closeStreamServer() {}
+
+// stub implementation when gostream not available
+func (svc *webService) initStreamServer(ctx context.Context, options *weboptions.Options) error {
+	return nil
+}
+
+// stub for missing gostream
+type options struct{}
+
 // stub for missing graphviz
 func (svc *webService) handleVisualizeResourceGraph(w http.ResponseWriter, r *http.Request) {}
-
-func (svc *webService) makeStreamServer(ctx context.Context) (*StreamServer, error) {
-	svc.refreshVideoSources()
-	var streams []gostream.Stream
-	var streamTypes []bool
-
-	if svc.opts.streamConfig == nil || len(svc.videoSources) == 0 {
-		if len(svc.videoSources) != 0 {
-			svc.logger.Debug("not starting streams due to no stream config being set")
-		}
-		noopServer, err := gostream.NewStreamServer(streams...)
-		return &StreamServer{noopServer, false}, err
-	}
-
-	addStream := func(streams []gostream.Stream, name string, isVideo bool) ([]gostream.Stream, error) {
-		config := *svc.opts.streamConfig
-		config.Name = name
-		if isVideo {
-			config.AudioEncoderFactory = nil
-
-			// set TargetFrameRate to the framerate of the video source if available
-			props, err := svc.videoSources[name].MediaProperties(ctx)
-			if err != nil {
-				svc.logger.Warnw("failed to get video source properties", "name", name, "error", err)
-			} else if props.FrameRate > 0.0 {
-				// round float up to nearest int
-				config.TargetFrameRate = int(math.Ceil(float64(props.FrameRate)))
-			}
-			// default to 60fps if the video source doesn't have a framerate
-			if config.TargetFrameRate == 0 {
-				config.TargetFrameRate = 60
-			}
-
-			if runtime.GOOS == "windows" {
-				// TODO(RSDK-1771): support video on windows
-				svc.logger.Warnw("not starting video stream since not supported on Windows yet", "name", name)
-				return streams, nil
-			}
-		} else {
-			config.VideoEncoderFactory = nil
-		}
-		stream, err := gostream.NewStream(config)
-		if err != nil {
-			return streams, err
-		}
-		return append(streams, stream), nil
-	}
-	for name := range svc.videoSources {
-		var err error
-		streams, err = addStream(streams, name, true)
-		if err != nil {
-			return nil, err
-		}
-		streamTypes = append(streamTypes, true)
-	}
-
-	streamServer, err := gostream.NewStreamServer(streams...)
-	if err != nil {
-		return nil, err
-	}
-
-	for idx, stream := range streams {
-		if streamTypes[idx] {
-			svc.startVideoStream(ctx, svc.videoSources[stream.Name()], stream)
-		}
-	}
-
-	return &StreamServer{streamServer, true}, nil
-}
