+++ robot/web/web_options_cgo.go
@@ -1,3 +1,5 @@
+//go:build !no_cgo
+
 package web
 
 import "go.viam.com/rdk/gostream"
