+++ web/server/entrypoint_cgo.go
@@ -1,3 +1,5 @@
+//go:build !no_cgo
+
 package server
 
 import (
